requirements.txt
Flask==2.3.2
pydub==0.25.1
ffmpeg-python==0.2.0
librosa==0.10.1
python-pptx
